requirements_cpu.txt
gin-config>=0.3.0
paddlepaddle==1.8.3